feat(web): improvements & fixes
