+++ apps/dashboard/src/components/primitives/sonner-helpers.tsx
@@ -1,5 +1,5 @@
 import { ExternalToast, toast } from 'sonner';
-import { PromoteToast, Toast, ToastProps } from './sonner';
+import { Toast, ToastProps } from './sonner';
 import { ReactNode } from 'react';
 
 export const showToast = ({
@@ -14,23 +14,6 @@ export const showToast = ({
     duration: 5000,
     unstyled: true,
     closeButton: false,
-    ...options,
-  });
-};
-
-export const promoteToast = ({
-  title,
-  description,
-  action,
-  options,
-}: {
-  title: string;
-  description: string;
-  action: { label: string; onClick: () => void };
-  options?: ExternalToast;
-}) => {
-  return toast.custom((t) => <PromoteToast t={t} title={title} description={description} action={action} />, {
-    duration: 5000,
     ...options,
   });
 };
